requirements.txt
watchdog
pyyaml
sentence-transformers
scikit-learn
langchain-core
langchain-community
langchain-openai
openai
python-dotenv
duckduckgo-search
tqdm
